refactor: Update API response model
